requirements.txt

astroid==2.11.2
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cleo==0.8.1
click==8.1.0
clikit==0.6.2
crashtest==0.3.1
DateTime==4.4
dill==0.3.4
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
filelock==3.4.2
Flask==2.1.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
icalendar==4.0.9
icalevents==0.1.25
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
jicson==1.0.2
Jinja2==3.1.1
keyring==21.8.0
lazy-object-proxy==1.7.1
libsass==0.21.0
lockfile==0.12.2
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.1
pluggy==1.0.0
poetry==1.1.12
poetry-core==1.0.7
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pylev==1.4.0
pylint==2.13.3
PyMySQL==1.0.2
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.4.32
toml==0.10.2
tomli==2.0.1
tomlkit==0.8.0
urllib3==1.26.8
virtualenv==20.13.0
webencodings==0.5.1
Werkzeug~=2.0.0
wrapt==1.14.0
WTForms==3.0.1
zope.interface==5.4.0
